requirements.txt
ifcopenshell
psycopg2-binary
python-dotenv
antrhopic